code/deployment/api/requirements.txt
fastapi==0.95.2
uvicorn==0.18.3
Pillow==9.2.0
numpy==1.23.0
torch==2.2.1
torchvision==0.17.1
python-multipart